Load Global Country Boundaries

Load Global Country Boundaries

Usage

load_world_country(center = "west")

Arguments

center

(optional) Center must be west or east. Default is west.

Value

An sf object

Details

When the center parameter is set to west, the map center is the Atlantic Ocean; and when center is east, the map center is the Pacific Ocean.

Examples

load_world_country()
#> Simple feature collection with 249 features and 39 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 83.6341
#> Geodetic CRS:  WGS 84
#> # A tibble: 249 × 40
#>    country1     country2 capital1 capital2 scalerank labelrank sovereignt sov_a3
#>    <chr>        <chr>    <chr>    <chr>        <int>     <dbl> <chr>      <chr> 
#>  1 阿鲁巴       Aruba    奥拉涅…  Oranjes…         3         5 Netherlan… NL1   
#>  2 阿富汗       Afghani… 喀布尔   Kabul            0         3 Afghanist… AFG   
#>  3 安哥拉       Angola   罗安达   Luanda           0         3 Angola     AGO   
#>  4 安圭拉       Anguilla 瓦利     The Val…         3         6 United Ki… GB1   
#>  5 阿尔巴尼亚   Albania  地拉那   Tirana           0         6 Albania    ALB   
#>  6 奥兰群岛     Aland I… 玛丽港   Marieha…         3         6 Finland    FI1   
#>  7 安道尔共和国 Andorra  安道尔   Andorra…         0         6 Andorra    AND   
#>  8 阿拉伯联合…  United … 阿布扎比 Abu Dha…         0         4 United Ar… ARE   
#>  9 阿根廷       Argenti… 布宜诺…  BuenosA…         0         2 Argentina  ARG   
#> 10 亚美尼亚     Armenia  埃里温   Yerevan          0         6 Armenia    ARM   
#> # ℹ 239 more rows
#> # ℹ 32 more variables: type <chr>, name <chr>, name_long <chr>, abbrev <chr>,
#> #   postal <chr>, formal_en <chr>, formal_fr <chr>, note_adm0 <chr>,
#> #   note_brk <chr>, name_sort <chr>, name_alt <chr>, mapcolor7 <dbl>,
#> #   mapcolor8 <dbl>, mapcolor9 <dbl>, mapcolor13 <dbl>, pop_est <dbl>,
#> #   gdp_md_est <dbl>, pop_year <dbl>, lastcensus <dbl>, gdp_year <dbl>,
#> #   economy <chr>, income_grp <chr>, iso_a2 <chr>, iso_n3 <chr>, un_a3 <chr>, …